### 2023-02-13 #573 @bnicenboim
presubmission inquiry for pangoling: Access to word predictability using large language (transformer) models.
Submitting Author Name: Bruno Nicenboim
Submitting Author Github Handle: @bnicenboim
Repository:  https://github.com/bnicenboim/pangoling
Submission type: Pre-submission

---

```
Package: pangoling
Title: Access to Large Language Model Predictions
Version: 0.0.0.9002
    person("Bruno", "Nicenboim",
    email = "bruno.nicenboim@gmail.com",
    role = c( "aut","cre"),
    person("Chris", "Emmerly", role = "ctb"),
    person("Giovanni", "Cassani", role = "ctb"))
Description: Access to word predictability using large language (transformer) models.
LazyData: false
Config/reticulate:
  list(
    packages = list(
      list(package = "torch"),
      list(package = "transformers")
    )
  )
    memoise,
    reticulate,
    tidyselect,
    tidytable (>= 0.7.2),
    cachem
    tictoc,
    covr,
    spelling
StagedInstall: yes

```

The package is a wrapper around [`transformers`](https://pypi.org/project/transformers/)  python package, and it can tokenize, get word predictability and calculate perplexity which is  text analysis.

NA

This is mostly for psycho/neuro/- linguists that use word predictability as a predictor in their research, such as ERP and reading.

Another R package that acts as a wrapper for [`transformers`](https://pypi.org/project/transformers/) is [`text`](https://r-text.org//) However, `text` is more general, and its focus
is on Natural Language Processing and Machine Learning. `pangoling` is much more specific and the focus is on measures used as predictors in analyses of data from experiments, rather than NLP.

Yes, the output of pkgcheck  fails only because of the use of `<<-`. But this is done in order to use memoise, as it is recommended in [its page](https://memoise.r-lib.org/reference/memoise.html). The `<<-` in the package appears inside `.onLoad <- function(libname, pkgname) {`

```
 # caching:
  tokenizer <<- memoise::memoise(tokenizer)
  lang_model <<- memoise::memoise(lang_model)
  transformer_vocab <<- memoise::memoise(transformer_vocab)
```

The pkgcheck output is the following:

── pangoling 0.0.0.9002 ────────────────────────────

✔ Package name is available
✔ has a 'codemeta.json' file.
✔ has a 'contributing' file.
✔ uses 'roxygen2'.
✔ 'DESCRIPTION' has a URL field.
✔ 'DESCRIPTION' has a BugReports field.
✔ Package has at least one HTML vignette
✔ All functions have examples.
✖ Package uses global assignment operator ('<<-').
✔ Package has continuous integration checks.
✔ Package coverage is 94.4%.
✔ R CMD check found no errors.
✔ R CMD check found no warnings.

ℹ Current status:
✖ This package is not ready to be submitted.
