### 2023-02-15 #573 @maurolepore
Dear @bnicenboim,

It's my first rotation as EiC and I'm still learning the nuances of assessing the eligibility of each submission. I need to show other editors the evidence that shows how pangoling meets our criteria for fit and overlap. Can you please help me by answering the following questions as succinctly and precisely as possible? I numbered the actionable items to help you respond to them specifically.

#### [Package categories](https://devguide.ropensci.org/policies.html#package-categories)

scientific software wrappers: Packages that wrap non-R utility programs used for scientific research.

- [ ] ml01. Must be specific to research fields, not general computing utilities.

What research field is pangoling specific for? Is that psycho/neuro/- linguists?

- [ ] ml02. Must be non-trivial.

Can you please expand on what value pangoling adds? That is, above simple [system()](https://rdrr.io/r/base/system.html) call or bindings, whether in parsing inputs and outputs, data handling, etc. Improved installation process, or extension of compatibility to more platforms, may constitute added value if installation is complex.

#### [Other scope considerationos](https://devguide.ropensci.org/policies.html#other-scope-considerations)

- [ ] ml03. Should be general in the sense that they should solve a problem as broadly as possible while maintaining a coherent user interface and code base. For instance, if several data sources use an identical API, we prefer a package that provides access to all the data sources, rather than just one.

The 'pangoling' package states that the overlaping package 'text' is more general. Can you please argue how despite this pangoling still meets rOpenSci's guidelines for fit and overlap?

#### [Package overlap](https://devguide.ropensci.org/policies.html#overlap)

- [ ] ml04. Avoids duplication of functionality of existing R packages in any repo without significant improvements.

Can you please explain if 'pangoling' duplicates or not functionaliry in the 'text' or another R package, and if it does then how 'pangoling' represents a significant improvement (see our guide for details on what we mean by 'significant improvement')?

- [ ] ml05. Also, as it becomes increasingly easy to call python packages from R, can you please explain how straightforward it would be to access the underlying python functionality without 'pangoling'?

Thanks for your patience :-)
