### 2023-02-15 #573 @bnicenboim
Ok, sure I answer inline.

> Dear @bnicenboim,
>
> It's my first rotation as EiC and I'm still learning the nuances of assessing the eligibility of each submission. I need to show other editors the evidence that shows how pangoling meets our criteria for fit and overlap. Can you please help me by answering the following questions as succinctly and precisely as possible? I numbered the actionable items to help you respond to them specifically.
> #### [Package categories](https://devguide.ropensci.org/policies.html#package-categories)
>
> scientific software wrappers: Packages that wrap non-R utility programs used for scientific research.
>
>     * [x]  ml01. Must be specific to research fields, not general computing utilities.
>
>
> What research field is pangoling specific for? Is that psycho/neuro/- linguists?

Yes, it's common to use word predictability as a predictor in models, and pangoling extract predictability from the transformer models.

>
>     * [ ]  ml02. Must be non-trivial.
>
>
> Can you please expand on what value pangoling adds? That is, above simple [system()](https://rdrr.io/r/base/system.html) call or bindings, whether in parsing inputs and outputs, data handling, etc. Improved installation process, or extension of compatibility to more platforms, may constitute added value if installation is complex.

Transformer models are "meant" to be used for computational linguistic tasks. For example gpt-like models produce a (random) continuation given a context. That's trivial to get, since there is a short-cut call called `pipeline()` in python that does exactly that. The thing is that one can also get the probability of each word in a given text without generating anything, that's less trivial to obtain, but it's very useful in *-linguistics. It's less trivial because one needs to know how to set up the language model, then one obtains a huge tensor (which is not trivial to manipulate for most R users) and finally one needs to take care of the mapping between words and phrases (the important thing in *ling) and the tokens (which is how the model is encoding the words), and this correspondence might be one to one or one to many. For Bert-like model the challenges are similar. Crucially one needs to understand how these large language transformer models work. There are two contributors of the package, that was their role, explaining me how these models work so that I could figure out what were the python functions I needed :)

Also the point of using `memoise` is that the package is not object oriented (like R6) which are more confusing for most basic R users, and completely functional. It just remembers what was the last type of language model that was used.

I hope it's clearer, but feel free to ask!

> #### [Other scope considerationos](https://devguide.ropensci.org/policies.html#other-scope-considerations)
>
>     * [ ]  ml03. Should be general in the sense that they should solve a problem as broadly as possible while maintaining a coherent user interface and code base. For instance, if several data sources use an identical API, we prefer a package that provides access to all the data sources, rather than just one.
>
>
> The 'pangoling' package states that the overlaping package 'text' is more general. Can you please argue how despite this pangoling still meets rOpenSci's guidelines for fit and overlap?

`text` brings `transformers` python package to R, and adds some machine learning stuff.  I would say that also here, the overlap is that `text` is more general and it doesn't allow for generating `pangoling` output in a straightforward way. In fact, I'm not sure if it's even possible since it seems more limited than `transformers`.
I would say that the users of pangoling would be mostly psycho/neuro linguists, while the users of text are computational linguists.

> #### [Package overlap](https://devguide.ropensci.org/policies.html#overlap)
>
>     * [ ]  ml04. Avoids duplication of functionality of existing R packages in any repo without significant improvements.
>
>
> Can you please explain if 'pangoling' duplicates or not functionaliry in the 'text' or another R package, and if it does then how 'pangoling' represents a significant improvement (see our guide for details on what we mean by 'significant improvement')?

I think I answered in the previous point. I'm not even sure that you can get the output of pangoling just using `text`. I think the overlap is that they are both wrappers of `transformers`.

>
>     * [ ]  ml05. Also, as it becomes increasingly easy to call python packages from R, can you please explain how straightforward it would be to access the underlying python functionality without 'pangoling'?

I'm not sure if I understand this. One would need to set up the models in python, then extract the tensors and manipulate them. Finally, one needs to take care of the mapping between words and tokens. But python is not needed here.

>
>
> Thanks for your patience :-)

ok, there was a lot of overlap in my answers, so feel free to ask me more specific questions if something is not clear.
