### 2023-02-16 00b3d3c
bug in causal_tokens_lp_tbl
