### 2023-02-16 0ef4500
minor change in NEWS
