+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   intro-bert: intro-bert.html
   intro-gpt2: intro-gpt2.html
-last_built: 2023-02-13T22:26Z
+last_built: 2023-02-16T16:21Z
 urls:
   reference: http://bruno.nicenboim.me/pangoling/reference
   article: http://bruno.nicenboim.me/pangoling/articles
