### 2023-02-17 d8db102
causal_tokens_lp_tbl accepts batches but does sth wrong
