+++ R/tr_utils.R
@@ -100,14 +100,14 @@ get_vocab <- function(tkzr) {
 }
 
 encode <- function(x, tkzr, add_special_tokens = NULL, ...) {
-  if (!is.null(add_special_tokens)) {
-    tkzr$encode(x,
-      return_tensors = "pt",
-      add_special_tokens = add_special_tokens, ...
-    )
-  } else {
-    tkzr$encode(x, return_tensors = "pt", ...)
-  }
+    if (!is.null(add_special_tokens)) {
+      tkzr$batch_encode_plus(x,
+                             return_tensors = "pt",
+                             add_special_tokens = add_special_tokens, ...
+      )
+    } else {
+      tkzr$batch_encode_plus(x, return_tensors = "pt", ...)
+    }
 }
 
 get_word_by_word_texts <- function(x, .by) {
@@ -253,6 +253,7 @@ create_tensor_lst <- function(texts,
     !is.null(tkzr$special_tokens_map$eos_token)) {
     tkzr$pad_token <- tkzr$eos_token
   }
+  texts <- unlist(texts)
   # If I runt the following line, some models such as
   # 'flax-community/gpt-2-spanish' give a weird error of
   # 'GPT2TokenizerFast' object has no attribute 'is_fast'
@@ -273,7 +274,7 @@ create_tensor_lst <- function(texts,
   text_ids <- tidytable::map2(c(1,cumsum(g_batches)[-length(g_batches)]), cumsum(g_batches),
                   ~ seq(.x,.y))
   lapply(text_ids, function(text_id) {
-    tensor <- encode(x = texts[text_id],
+    tensor <- encode(x = as.list(texts[text_id]),
                      tkzr = tkzr,
       add_special_tokens = add_special_tokens,
       stride = as.integer(stride),
