### 2023-02-17 #573 @maurolepore
@bnicenboim, I'm still discussing the scope with other editors.

- [ ] ml06. Did you consider submitting pangoling as as [stats package](https://stats-devguide.ropensci.org/)? If so, what convinced you to submit as a general package?
