### 2023-02-17 #573 @maurolepore
Thanks.

I ask because the category "text analysis" in the standard-package guide states:

> Machine-learning and packages implementing NLP analysis algorithms should be submitted under [statistical software peer review](https://stats-devguide.ropensci.org/).

Knowing that you at least considered it I can now be sure the standard-package review is your informed decision.
