### 2023-02-17 #573 @bnicenboim
Thanks for the clarification, i found this under statistical software:

   1. Bayesian and Monte Carlo Routines
   2. Dimensionality Reduction, Clustering, and Unsupervised Learning
   3. Machine Learning
   4. Regression and Supervised Learning
   5. Exploratory Data Analysis (EDA) and Summary Statistics
   6. Spatial Analyses
   7. Time Series Analyses

And the package doesn't fall into any of those, it's not doing machine
learning either. So I think it's fine under general.
