### 2023-02-18 d01d502
batches for causal models work
