+++ R/tr_causal.R
@@ -95,6 +95,7 @@ causal_next_tokens_tbl <- function(context,
                                    add_special_tokens = NULL,
                                    config_model = NULL,
                                    config_tokenizer = NULL) {
+  if(length(unlist(context)) > 1) stop2("Only one context is allowed in this function.")
   message_verbose("Processing using causal model '", model, "'...")
   trf <- lang_model(model,
     task = "causal",
@@ -106,10 +107,11 @@ causal_next_tokens_tbl <- function(context,
   )
 
   lang_model(model, task = "causal", config_model = config_model)
-  context_tensor <- encode(context,
+  #no batches allowed
+  context_tensor <- encode(list(unlist(context)),
     tkzr,
     add_special_tokens = add_special_tokens
-  )
+  )$input_ids
   generated_outputs <- trf(context_tensor)
   n_tokens <- length(context_tensor$tolist()[0])
   logits_next_word <- generated_outputs$logits[0][n_tokens - 1]
@@ -132,7 +134,11 @@ causal_next_tokens_tbl <- function(context,
 #' @param x Vector of words, phrases or texts.
 #' @param .by Vector that indicates how the text should be split.
 #' @inheritParams causal_preload
-#' @param ignore_regex Can ignore certain characters when calculates the log probabilities. For example `^[[:punct:]]$` will ignore all punctuation  that stands alone in a token.
+#' @param ignore_regex Can ignore certain characters when calculates the log
+#'                      probabilities. For example `^[[:punct:]]$` will ignore
+#'                      all punctuation  that stands alone in a token.
+#' @param batch_size Maximum size of the batch. Larges batches speedup
+#'                   processing but take more memory.
 #' @inherit  causal_preload details
 #' @inheritSection causal_next_tokens_tbl More examples
 #' @return A named vector of log probabilities.
@@ -151,7 +157,8 @@ causal_lp <- function(x,
                       model = getOption("pangoling.causal.default"),
                       add_special_tokens = NULL,
                       config_model = NULL,
-                      config_tokenizer = NULL) {
+                      config_tokenizer = NULL,
+                      batch_size = 1) {
   stride <- 1 # fixed for now
   if (length(x) <= 1) stop2("The argument `x` needs at least two elements.")
   message_verbose("Processing using causal model '", model, "'...")
@@ -170,26 +177,31 @@ causal_lp <- function(x,
     config_model = config_model
   )
   tensors <- create_tensor_lst(
-    texts = pasted_texts,
+    texts = unname(pasted_texts),
     tkzr = tkzr,
     add_special_tokens = add_special_tokens,
-    stride = stride
+    stride = stride,
+    batch_size = batch_size
   )
-  out <- tidytable::pmap.(
+
+  lmats <- lapply(tensors, function(tensor) causal_mat(tensor,
+                     trf,
+                     tkzr,
+                     add_special_tokens = add_special_tokens,
+                     stride = stride
+  )) |>
+    unlist(recursive = FALSE)
+    out <- tidytable::pmap.(
     list(
       word_by_word_texts,
-      names(word_by_word_texts), tensors
+      names(word_by_word_texts),
+      lmats
     ),
-    function(words, item, tensor) {
+    function(words, item, mat) {
       # words <- word_by_word_texts[[1]]
       # item <- names(texts[1])
-      # tensor <- tensors[[1]]
-      mat <- causal_mat(tensor,
-        trf,
-        tkzr,
-        add_special_tokens = add_special_tokens,
-        stride = stride
-      )
+      # mat <- lmats[[1]]
+
       message_verbose(
         "Text id: ", item, "\n`",
[…]
