+++ NEWS.md
@@ -9,5 +9,8 @@
 # pangoling 0.0.0.9002
 * minor function names to avoid conflict with other packages
 
-# pangoling 0.0.0.9002
+# pangoling 0.0.0.9003
 * bug in causal_tokens_lp_tbl fixed
+
+# pangoling 0.0.0.9004
+* Causal models accept batches. 
