### 2023-02-19 74f368b
Accepts strings with no tokens
