+++ R/tr_causal.R
@@ -95,7 +95,7 @@ causal_next_tokens_tbl <- function(context,
                                    add_special_tokens = NULL,
                                    config_model = NULL,
                                    config_tokenizer = NULL) {
-  if(length(unlist(context)) > 1) stop2("Only one context is allowed in this function.")
+  if (length(unlist(context)) > 1) stop2("Only one context is allowed in this function.")
   message_verbose("Processing using causal model '", model, "'...")
   trf <- lang_model(model,
     task = "causal",
@@ -107,7 +107,7 @@ causal_next_tokens_tbl <- function(context,
   )
 
   lang_model(model, task = "causal", config_model = config_model)
-  #no batches allowed
+  # no batches allowed
   context_tensor <- encode(list(unlist(context)),
     tkzr,
     add_special_tokens = add_special_tokens
@@ -160,7 +160,6 @@ causal_lp <- function(x,
                       config_tokenizer = NULL,
                       batch_size = 1) {
   stride <- 1 # fixed for now
-  if (length(x) <= 1) stop2("The argument `x` needs at least two elements.")
   message_verbose("Processing using causal model '", model, "'...")
   word_by_word_texts <- get_word_by_word_texts(x, .by)
 
@@ -184,14 +183,16 @@ causal_lp <- function(x,
     batch_size = batch_size
   )
 
-  lmats <- lapply(tensors, function(tensor) causal_mat(tensor,
-                     trf,
-                     tkzr,
-                     add_special_tokens = add_special_tokens,
-                     stride = stride
-  )) |>
+  lmats <- lapply(tensors, function(tensor) {
+    causal_mat(tensor,
+      trf,
+      tkzr,
+      add_special_tokens = add_special_tokens,
+      stride = stride
+    )
+  }) |>
     unlist(recursive = FALSE)
-    out <- tidytable::pmap.(
+  out <- tidytable::pmap.(
     list(
       word_by_word_texts,
       names(word_by_word_texts),
@@ -276,11 +277,18 @@ causal_tokens_lp_tbl <- function(texts,
   }) |>
     unlist(recursive = FALSE)
 
-  tidytable::map_dfr.(ls_mat, function( mat) {
-    tidytable::tidytable(
-      token = colnames(mat),
-      lp = tidytable::map2_dbl.(colnames(mat), seq_len(ncol(mat)), ~ mat[.x, .y])
-    )
+  tidytable::map_dfr.(ls_mat, function(mat) {
+    if (ncol(mat) ==1 && colnames(mat) == "") {
+      tidytable::tidytable(
+        token = "",
+        lp = NA_real_
+      )
+    } else {
+      tidytable::tidytable(
+        token = colnames(mat),
+        lp = tidytable::map2_dbl.(colnames(mat), seq_len(ncol(mat)), ~ mat[.x, .y])
+      )
+    }
   }, .id = .id)
 }
 
@@ -291,48 +299,64 @@ causal_mat <- function(tensor,
                        tkzr,
                        add_special_tokens = NULL,
                        stride = 1) {
-
   message_verbose(
     "Processing a batch of size ",
     tensor$input_ids$shape[0],
     " with ",
     tensor$input_ids$shape[1], " tokens."
   )
- #logits_b <- trf(tensor)$logits
 
-  logits_b <- trf$forward(input_ids = tensor$input_ids,
[…]
