+++ DESCRIPTION
@@ -41,7 +41,7 @@ Config/testthat/edition: 3
 RoxygenNote: 7.2.3
 Roxygen: list(markdown = TRUE)
 Depends: 
-    R (>= 2.10)
+    R (>= 4.1.0)
 VignetteBuilder: knitr
 StagedInstall: yes
 Language: en-US
