### 2023-02-21 #573 @bnicenboim
Thanks, should I do something about `Package uses global assignment operator ('<<-').`. If I add pkgcheck in github actions it will just fail.
