+++ R/tr_causal.R
@@ -278,7 +278,7 @@ causal_tokens_lp_tbl <- function(texts,
     unlist(recursive = FALSE)
 
   tidytable::map_dfr.(ls_mat, function(mat) {
-    if (ncol(mat) ==1 && colnames(mat) == "") {
+    if (ncol(mat) == 1 && colnames(mat) == "") {
       tidytable::tidytable(
         token = "",
         lp = NA_real_
