+++ tests/testthat/test-tr_causal.R
@@ -25,21 +25,21 @@ test_that("empty or small strings", {
   expect_equal(as.data.frame(lp_it), data.frame(token = "It", lp = NA_real_))
   expect_warning(lp_NA <- causal_tokens_lp_tbl(texts = ""))
   expect_equal(as.data.frame(lp_NA), data.frame(token = "", lp = NA_real_))
-  small_str <- c("It", "It","is")
-  lp_small <- causal_lp(x =small_str ,.by = c(1,2,2))
+  small_str <- c("It", "It", "is")
+  lp_small <- causal_lp(x = small_str, .by = c(1, 2, 2))
   expect_equal(lp_small[1:2], c(It = NA_real_, It = NA_real_))
-  expect_warning(lp_small <- causal_lp(x =c("","It"), .by = c(1,2)))
+  expect_warning(lp_small <- causal_lp(x = c("", "It"), .by = c(1, 2)))
   expect_equal(lp_small, c(NA_real_, "It" = NA_real_))
 })
 
 
 test_that("long input work", {
   skip_if_no_python_stuff()
-  long0 <- paste(rep("x",1022), collapse =" ")
-  long <- paste(rep("x",1024), collapse =" ")
-  longer <- paste(rep("x",1025), collapse =" ")
-  lp_long0 <- causal_tokens_lp_tbl(c(long0, long, longer),add_special_tokens = TRUE, batch_size = 3, model = "sshleifer/tiny-gpt2")
-  lp_long1 <- causal_tokens_lp_tbl(c(long0, long, longer),add_special_tokens = TRUE, batch_size = 1, model = "sshleifer/tiny-gpt2")
+  long0 <- paste(rep("x", 1022), collapse = " ")
+  long <- paste(rep("x", 1024), collapse = " ")
+  longer <- paste(rep("x", 1025), collapse = " ")
+  lp_long0 <- causal_tokens_lp_tbl(c(long0, long, longer), add_special_tokens = TRUE, batch_size = 3, model = "sshleifer/tiny-gpt2")
+  lp_long1 <- causal_tokens_lp_tbl(c(long0, long, longer), add_special_tokens = TRUE, batch_size = 1, model = "sshleifer/tiny-gpt2")
   expect_equal(lp_long0, lp_long1)
 })
 
