### 2023-02-21 #573 @mpadge
@bnicenboim @maurolepore I've just updated `pkgcheck` via the issue linked above to allow `<<-` in an `.onLoad` function for use of memoise. The {pangoling} package still fails because it also has two `.onLoad` entrys for `<<- reticulate::import`. This is also recommended in the reticulate package. Although not yet permitted in `pkgcheck`, this use for reticulate imports will also be permitted soon, and the pangoling package will then pass all tests. In the meantime, @bnicenboim please simply add an explanatory note, and links to this comment or the pkgcheck issue as you see fit. thanks
